refactor: Move files for explicit folders
