+++ src/containers/SidebarActions/NotificationsModal.tsx
@@ -1,9 +1,13 @@
 import {useMemo, type FC} from "react"
 import {IoCloseCircle} from "react-icons/io5"
-import Button, {ButtonColor, ButtonSize, ButtonVariant} from "../Button"
-import Typography, {TypographyVariant} from "../Typography"
+import Button, {
+  ButtonColor,
+  ButtonSize,
+  ButtonVariant,
+} from "../../components/Button"
+import Typography, {TypographyVariant} from "../../components/Typography"
 import {twMerge} from "tailwind-merge"
-import IconButton from "../IconButton"
+import IconButton from "../../components/IconButton"
 import {useSidebarModalActiveStore} from "./useSidebarActions"
 import useCachedNotifications, {
   useNotificationsStateStore,
